+++ resources/views/pages/main/demo/distribusi-widget.blade.php
@@ -105,8 +105,13 @@
 @endsection
 
 @php
-    $demoDir = resource_path('views/pages/demo');
-    $demoFiles = \Illuminate\Support\Facades\File::allFiles($demoDir);
+    $demoDir = resource_path('views/pages/main/demo');
+    if (!\Illuminate\Support\Facades\File::isDirectory($demoDir)) {
+        $demoDir = resource_path('views/pages/demo');
+    }
+    $demoFiles = \Illuminate\Support\Facades\File::isDirectory($demoDir)
+        ? \Illuminate\Support\Facades\File::allFiles($demoDir)
+        : [];
 
     $widgetsByCategory = [];
     $flexibleWidgetsMap = [];